Use 64-bit arithmetic for computing array offsets. (#1059)
* Use 64-bit arithmetic for computing array offsets.

This is not a complete solution to #134, because individual array
dimensions must still fit in a signed 32-bit integer.  However, it
does allow the product of dimensions to be large.

* Use 64-bit arithmetic for computing thread IDs.

This allows kernels to have more than 2**32-1 virtual threads,
although the number of groups must still be less than this.

Note that this is not about *physical* threads, where having this many
would be pointlessly large, but about virtual threads.

* It is apparently better to use zero-extension here.

* More tags; just stop running this!

* This also needs 64-bit indexes.

+++ src/Futhark/CodeGen/Backends/GenericC.hs
@@ -1873,7 +1873,7 @@ compileExp = compilePrimExp compileLeaf
           return [C.cexp|$id:src[$exp:iexp']|]
 
         compileLeaf (SizeOf t) =
-          return [C.cexp|(typename int32_t)sizeof($ty:t')|]
+          return [C.cexp|(typename int64_t)sizeof($ty:t')|]
           where t' = primTypeToCType t
 
 -- | Tell me how to compile a @v@, and I'll Compile any @PrimExp v@ for you.
